router: moved static methods to handlers
