+++ eclair-core/src/main/scala/fr/acinq/eclair/router/Router.scala
@@ -18,11 +18,11 @@ package fr.acinq.eclair.router
 
 import akka.Done
 import akka.actor.{ActorRef, Props}
+import akka.event.DiagnosticLoggingAdapter
 import akka.event.Logging.MDC
-import akka.event.{DiagnosticLoggingAdapter, LoggingAdapter}
 import fr.acinq.bitcoin.Crypto.PublicKey
 import fr.acinq.bitcoin.Script.{pay2wsh, write}
-import fr.acinq.bitcoin.{ByteVector32, ByteVector64, Satoshi}
+import fr.acinq.bitcoin.{ByteVector32, Satoshi}
 import fr.acinq.eclair.Logs.LogCategory
 import fr.acinq.eclair._
 import fr.acinq.eclair.blockchain._
@@ -31,23 +31,18 @@ import fr.acinq.eclair.crypto.TransportHandler
 import fr.acinq.eclair.db.NetworkDb
 import fr.acinq.eclair.io.Peer.PeerRoutingMessage
 import fr.acinq.eclair.payment.PaymentRequest.ExtraHop
-import fr.acinq.eclair.router.Graph.GraphStructure.DirectedGraph.graphEdgeToHop
-import fr.acinq.eclair.router.Graph.GraphStructure.{DirectedGraph, GraphEdge}
-import fr.acinq.eclair.router.Graph.{RichWeight, RoutingHeuristics, WeightRatios}
-import fr.acinq.eclair.router.Monitoring.{Metrics, Tags}
+import fr.acinq.eclair.router.Graph.GraphStructure.DirectedGraph
+import fr.acinq.eclair.router.Graph.WeightRatios
 import fr.acinq.eclair.transactions.Scripts
 import fr.acinq.eclair.wire._
 import kamon.context.Context
-import scodec.bits.ByteVector
-import shapeless.HNil
 
-import scala.annotation.tailrec
 import scala.collection.immutable.SortedMap
-import scala.collection.{SortedSet, mutable}
+import scala.collection.mutable
 import scala.compat.Platform
 import scala.concurrent.duration._
 import scala.concurrent.{ExecutionContext, Promise}
-import scala.util.{Random, Try}
+import scala.util.Try
 
 /**
  * Created by PM on 24/05/2016.
@@ -454,35 +449,7 @@ object Router {
   val shortChannelIdKey = Context.key[ShortChannelId]("shortChannelId", ShortChannelId(0))
   val remoteNodeIdKey = Context.key[String]("remoteNodeId", "unknown")
 
-  // maximum number of ids we can keep in a single chunk and still have an encoded reply that is smaller than 65Kb
-  // please note that:
-  // - this is based on the worst case scenario where peer want timestamps and checksums and the reply is not compressed
-  // - the maximum number of public channels in a single block so far is less than 300, and the maximum number of tx per block
-  // almost never exceeds 2800 so this is not a real limitation yet
-  val MAXIMUM_CHUNK_SIZE = 3200
-
   def props(nodeParams: NodeParams, watcher: ActorRef, initialized: Option[Promise[Done]] = None) = Props(new Router(nodeParams, watcher, initialized))
-
-  def toFakeUpdate(extraHop: ExtraHop, htlcMaximum: MilliSatoshi): ChannelUpdate = {
-    // the `direction` bit in flags will not be accurate but it doesn't matter because it is not used
-    // what matters is that the `disable` bit is 0 so that this update doesn't get filtered out
-    ChannelUpdate(signature = ByteVector64.Zeroes, chainHash = ByteVector32.Zeroes, extraHop.shortChannelId, Platform.currentTime.milliseconds.toSeconds, messageFlags = 1, channelFlags = 0, extraHop.cltvExpiryDelta, htlcMinimumMsat = 0 msat, extraHop.feeBase, extraHop.feeProportionalMillionths, Some(htlcMaximum))
-  }
-
-  def toAssistedChannels(extraRoute: Seq[ExtraHop], targetNodeId: PublicKey, amount: MilliSatoshi): Map[ShortChannelId, AssistedChannel] = {
-    // BOLT 11: "For each entry, the pubkey is the node ID of the start of the channel", and the last node is the destination
-    // The invoice doesn't explicitly specify the channel's htlcMaximumMsat, but we can safely assume that the channel
-    // should be able to route the payment, so we'll compute an htlcMaximumMsat accordingly.
-    // We could also get the channel capacity from the blockchain (since we have the shortChannelId) but that's more expensive.
-    // We also need to make sure the channel isn't excluded by our heuristics.
-    val lastChannelCapacity = amount.max(RoutingHeuristics.CAPACITY_CHANNEL_LOW)
-    val nextNodeIds = extraRoute.map(_.nodeId).drop(1) :+ targetNodeId
-    extraRoute.zip(nextNodeIds).reverse.foldLeft((lastChannelCapacity, Map.empty[ShortChannelId, AssistedChannel])) {
-      case ((amount, acs), (extraHop: ExtraHop, nextNodeId)) =>
-        val nextAmount = amount + nodeFee(extraHop.feeBase, extraHop.feeProportionalMillionths, amount)
-        (nextAmount, acs + (extraHop.shortChannelId -> AssistedChannel(extraHop, nextNodeId, nextAmount)))
-    }._2
-  }
 
   def getDesc(u: ChannelUpdate, channel: ChannelAnnouncement): ChannelDesc = {
     // the least significant bit tells us if it is node1 or node2
@@ -527,414 +494,4 @@ object Router {
   }
 
   def getStaleChannels(channels: Iterable[PublicChannel], currentBlockHeight: Long): Iterable[PublicChannel] = channels.filter(data => isStale(data.ann, data.update_1_opt, data.update_2_opt, currentBlockHeight))
-
-  /**
-   * Filters channels that we want to send to nodes asking for a channel range
-   */
-  def keep(firstBlockNum: Long, numberOfBlocks: Long, id: ShortChannelId): Boolean = {
-    val height = id.blockHeight
-    height >= firstBlockNum && height < (firstBlockNum + numberOfBlocks)
-  }
-
-  def shouldRequestUpdate(ourTimestamp: Long, ourChecksum: Long, theirTimestamp_opt: Option[Long], theirChecksum_opt: Option[Long]): Boolean = {
-    (theirTimestamp_opt, theirChecksum_opt) match {
-      case (Some(theirTimestamp), Some(theirChecksum)) =>
-        // we request their channel_update if all those conditions are met:
-        // - it is more recent than ours
-        // - it is different from ours, or it is the same but ours is about to be stale
-        // - it is not stale
-        val theirsIsMoreRecent = ourTimestamp < theirTimestamp
-        val areDifferent = ourChecksum != theirChecksum
-        val oursIsAlmostStale = isAlmostStale(ourTimestamp)
-        val theirsIsStale = isStale(theirTimestamp)
-        theirsIsMoreRecent && (areDifferent || oursIsAlmostStale) && !theirsIsStale
-      case (Some(theirTimestamp), None) =>
-        // if we only have their timestamp, we request their channel_update if theirs is more recent than ours
-        val theirsIsMoreRecent = ourTimestamp < theirTimestamp
-        val theirsIsStale = isStale(theirTimestamp)
-        theirsIsMoreRecent && !theirsIsStale
-      case (None, Some(theirChecksum)) =>
-        // if we only have their checksum, we request their channel_update if it is different from ours
-        // NB: a zero checksum means that they don't have the data
-        val areDifferent = theirChecksum != 0 && ourChecksum != theirChecksum
-        areDifferent
-      case (None, None) =>
-        // if we have neither their timestamp nor their checksum we request their channel_update
-        true
-    }
-  }
-
-  def computeFlag(channels: SortedMap[ShortChannelId, PublicChannel])(
-    shortChannelId: ShortChannelId,
-    theirTimestamps_opt: Option[ReplyChannelRangeTlv.Timestamps],
-    theirChecksums_opt: Option[ReplyChannelRangeTlv.Checksums],
-    includeNodeAnnouncements: Boolean): Long = {
-    import QueryShortChannelIdsTlv.QueryFlagType._
-
-    val flagsNodes = if (includeNodeAnnouncements) INCLUDE_NODE_ANNOUNCEMENT_1 | INCLUDE_NODE_ANNOUNCEMENT_2 else 0
-
-    val flags = if (!channels.contains(shortChannelId)) {
-      INCLUDE_CHANNEL_ANNOUNCEMENT | INCLUDE_CHANNEL_UPDATE_1 | INCLUDE_CHANNEL_UPDATE_2
-    } else {
-      // we already know this channel
-      val (ourTimestamps, ourChecksums) = Router.getChannelDigestInfo(channels)(shortChannelId)
-      // if they don't provide timestamps or checksums, we set appropriate default values:
-      // - we assume their timestamp is more recent than ours by setting timestamp = Long.MaxValue
-      // - we assume their update is different from ours by setting checkum = Long.MaxValue (NB: our default value for checksum is 0)
-      val shouldRequestUpdate1 = shouldRequestUpdate(ourTimestamps.timestamp1, ourChecksums.checksum1, theirTimestamps_opt.map(_.timestamp1), theirChecksums_opt.map(_.checksum1))
-      val shouldRequestUpdate2 = shouldRequestUpdate(ourTimestamps.timestamp2, ourChecksums.checksum2, theirTimestamps_opt.map(_.timestamp2), theirChecksums_opt.map(_.checksum2))
-      val flagUpdate1 = if (shouldRequestUpdate1) INCLUDE_CHANNEL_UPDATE_1 else 0
-      val flagUpdate2 = if (shouldRequestUpdate2) INCLUDE_CHANNEL_UPDATE_2 else 0
-      flagUpdate1 | flagUpdate2
-    }
-
-    if (flags == 0) 0 else flags | flagsNodes
-  }
-
-  /**
-   * Handle a query message, which includes a list of channel ids and flags.
-   *
-   * @param nodes     node id -> node announcement
-   * @param channels  channel id -> channel announcement + updates
-   * @param ids       list of channel ids
-   * @param flags     list of query flags, either empty one flag per channel id
-   * @param onChannel called when a channel announcement matches (i.e. its bit is set in the query flag and we have it)
-   * @param onUpdate  called when a channel update matches
-   * @param onNode    called when a node announcement matches
-   *
-   */
-  def processChannelQuery(nodes: Map[PublicKey, NodeAnnouncement],
-                          channels: SortedMap[ShortChannelId, PublicChannel])(
-                           ids: List[ShortChannelId],
-                           flags: List[Long],
-                           onChannel: ChannelAnnouncement => Unit,
-                           onUpdate: ChannelUpdate => Unit,
-                           onNode: NodeAnnouncement => Unit)(implicit log: LoggingAdapter): Unit = {
-    import QueryShortChannelIdsTlv.QueryFlagType
-
-    // we loop over channel ids and query flag. We track node Ids for node announcement
-    // we've already sent to avoid sending them multiple times, as requested by the BOLTs
-    @tailrec
-    def loop(ids: List[ShortChannelId], flags: List[Long], numca: Int = 0, numcu: Int = 0, nodesSent: Set[PublicKey] = Set.empty[PublicKey]): (Int, Int, Int) = ids match {
-      case Nil => (numca, numcu, nodesSent.size)
-      case head :: tail if !channels.contains(head) =>
-        log.warning("received query for shortChannelId={} that we don't have", head)
-        loop(tail, flags.drop(1), numca, numcu, nodesSent)
-      case head :: tail =>
-        val numca1 = numca
-        val numcu1 = numcu
-        var sent1 = nodesSent
-        val pc = channels(head)
-        val flag_opt = flags.headOption
-        // no flag means send everything
-
-        val includeChannel = flag_opt.forall(QueryFlagType.includeChannelAnnouncement)
-        val includeUpdate1 = flag_opt.forall(QueryFlagType.includeUpdate1)
-        val includeUpdate2 = flag_opt.forall(QueryFlagType.includeUpdate2)
-        val includeNode1 = flag_opt.forall(QueryFlagType.includeNodeAnnouncement1)
-        val includeNode2 = flag_opt.forall(QueryFlagType.includeNodeAnnouncement2)
-
-        if (includeChannel) {
-          onChannel(pc.ann)
-        }
-        if (includeUpdate1) {
-          pc.update_1_opt.foreach { u =>
-            onUpdate(u)
-          }
-        }
-        if (includeUpdate2) {
-          pc.update_2_opt.foreach { u =>
-            onUpdate(u)
-          }
-        }
-        if (includeNode1 && !sent1.contains(pc.ann.nodeId1)) {
-          nodes.get(pc.ann.nodeId1).foreach { n =>
-            onNode(n)
-            sent1 = sent1 + pc.ann.nodeId1
-          }
-        }
-        if (includeNode2 && !sent1.contains(pc.ann.nodeId2)) {
-          nodes.get(pc.ann.nodeId2).foreach { n =>
-            onNode(n)
-            sent1 = sent1 + pc.ann.nodeId2
-          }
-        }
-        loop(tail, flags.drop(1), numca1, numcu1, sent1)
-    }
-
-    loop(ids, flags)
-  }
-
-  /**
-   * Returns overall progress on synchronization
-   *
-   * @return a sync progress indicator (1 means fully synced)
-   */
-  def syncProgress(sync: Map[PublicKey, Sync]): SyncProgress = {
-    // NB: progress is in terms of requests, not individual channels
-    val (pending, total) = sync.foldLeft((0, 0)) {
-      case ((p, t), (_, sync)) => (p + sync.pending.size, t + sync.total)
-    }
-    if (total == 0) {
-      SyncProgress(1)
-    } else {
-      SyncProgress((total - pending) / (1.0 * total))
-    }
-  }
-
-  /**
-   * This method is used after a payment failed, and we want to exclude some nodes that we know are failing
-   */
-  def getIgnoredChannelDesc(channels: Map[ShortChannelId, PublicChannel], ignoreNodes: Set[PublicKey]): Iterable[ChannelDesc] = {
-    val desc = if (ignoreNodes.isEmpty) {
-      Iterable.empty[ChannelDesc]
-    } else {
-      // expensive, but node blacklisting shouldn't happen often
-      channels.values
-        .filter(channelData => ignoreNodes.contains(channelData.ann.nodeId1) || ignoreNodes.contains(channelData.ann.nodeId2))
-        .flatMap(channelData => Vector(ChannelDesc(channelData.ann.shortChannelId, channelData.ann.nodeId1, channelData.ann.nodeId2), ChannelDesc(channelData.ann.shortChannelId, channelData.ann.nodeId2, channelData.ann.nodeId1)))
-    }
-    desc
-  }
-
-  def getChannelDigestInfo(channels: SortedMap[ShortChannelId, PublicChannel])(shortChannelId: ShortChannelId): (ReplyChannelRangeTlv.Timestamps, ReplyChannelRangeTlv.Checksums) = {
-    val c = channels(shortChannelId)
-    val timestamp1 = c.update_1_opt.map(_.timestamp).getOrElse(0L)
-    val timestamp2 = c.update_2_opt.map(_.timestamp).getOrElse(0L)
-    val checksum1 = c.update_1_opt.map(getChecksum).getOrElse(0L)
-    val checksum2 = c.update_2_opt.map(getChecksum).getOrElse(0L)
-    (ReplyChannelRangeTlv.Timestamps(timestamp1 = timestamp1, timestamp2 = timestamp2), ReplyChannelRangeTlv.Checksums(checksum1 = checksum1, checksum2 = checksum2))
-  }
-
-  def crc32c(data: ByteVector): Long = {
-    import com.google.common.hash.Hashing
-    Hashing.crc32c().hashBytes(data.toArray).asInt() & 0xFFFFFFFFL
-  }
-
-  def getChecksum(u: ChannelUpdate): Long = {
-    import u._
-
-    val data = serializationResult(LightningMessageCodecs.channelUpdateChecksumCodec.encode(chainHash :: shortChannelId :: messageFlags :: channelFlags :: cltvExpiryDelta :: htlcMinimumMsat :: feeBaseMsat :: feeProportionalMillionths :: htlcMaximumMsat :: HNil))
-    crc32c(data)
-  }
-
-  case class ShortChannelIdsChunk(firstBlock: Long, numBlocks: Long, shortChannelIds: List[ShortChannelId]) {
-    /**
-     *
-     * @param maximumSize maximum size of the short channel ids list
-     * @return a chunk with at most `maximumSize` ids
-     */
-    def enforceMaximumSize(maximumSize: Int) = {
-      if (shortChannelIds.size <= maximumSize) this else {
-        // we use a random offset here, so even if shortChannelIds.size is much bigger than maximumSize (which should
-        // not happen) peers will eventually receive info about all channels in this chunk
-        val offset = Random.nextInt(shortChannelIds.size - maximumSize + 1)
-        this.copy(shortChannelIds = this.shortChannelIds.slice(offset, offset + maximumSize))
-      }
-    }
-  }
-
-  /**
-   * Split short channel ids into chunks, because otherwise message could be too big
-   * there could be several reply_channel_range messages for a single query, but we make sure that the returned
-   * chunks fully covers the [firstBlockNum, numberOfBlocks] range that was requested
-   *
-   * @param shortChannelIds       list of short channel ids to split
-   * @param firstBlockNum         first block height requested by our peers
-   * @param numberOfBlocks        number of blocks requested by our peer
-   * @param channelRangeChunkSize target chunk size. All ids that have the same block height will be grouped together, so
-   *                              returned chunks may still contain more than `channelRangeChunkSize` elements
-   * @return a list of short channel id chunks
-   */
-  def split(shortChannelIds: SortedSet[ShortChannelId], firstBlockNum: Long, numberOfBlocks: Long, channelRangeChunkSize: Int): List[ShortChannelIdsChunk] = {
-    // see BOLT7: MUST encode a short_channel_id for every open channel it knows in blocks first_blocknum to first_blocknum plus number_of_blocks minus one
-    val it = shortChannelIds.iterator.dropWhile(_.blockHeight < firstBlockNum).takeWhile(_.blockHeight < firstBlockNum + numberOfBlocks)
-    if (it.isEmpty) {
-      List(ShortChannelIdsChunk(firstBlockNum, numberOfBlocks, List.empty))
-    } else {
-      // we want to split ids in different chunks, with the following rules by order of priority
-      // ids that have the same block height must be grouped in the same chunk
-      // chunk should contain `channelRangeChunkSize` ids
-      @tailrec
-      def loop(currentChunk: List[ShortChannelId], acc: List[ShortChannelIdsChunk]): List[ShortChannelIdsChunk] = {
-        if (it.hasNext) {
-          val id = it.next()
-          val currentHeight = currentChunk.head.blockHeight
-          if (id.blockHeight == currentHeight)
-            loop(id :: currentChunk, acc) // same height => always add to the current chunk
-          else if (currentChunk.size < channelRangeChunkSize) // different height but we're under the size target => add to the current chunk
-          loop(id :: currentChunk, acc) // different height and over the size target => start a new chunk
-          else {
-            // we always prepend because it's more efficient so we have to reverse the current chunk
-            // for the first chunk, we make sure that we start at the request first block
-            // for the next chunks we start at the end of the range covered by the last chunk
-            val first = if (acc.isEmpty) firstBlockNum else acc.head.firstBlock + acc.head.numBlocks
-            val count = currentChunk.head.blockHeight - first + 1
-            loop(id :: Nil, ShortChannelIdsChunk(first, count, currentChunk.reverse) :: acc)
-          }
-        }
-        else {
-          // for the last chunk, we make sure that we cover the requested block range
-          val first = if (acc.isEmpty) firstBlockNum else acc.head.firstBlock + acc.head.numBlocks
-          val count = numberOfBlocks - first + firstBlockNum
-          (ShortChannelIdsChunk(first, count, currentChunk.reverse) :: acc).reverse
-        }
-      }
-
-      val first = it.next()
-      val chunks = loop(first :: Nil, Nil)
-
-      // make sure that all our chunks match our max size policy
-      enforceMaximumSize(chunks)
-    }
-  }
-
-  /**
-   * Enforce max-size constraints for each chunk
-   *
-   * @param chunks list of short channel id chunks
-   * @return a processed list of chunks
-   */
-  def enforceMaximumSize(chunks: List[ShortChannelIdsChunk]): List[ShortChannelIdsChunk] = chunks.map(_.enforceMaximumSize(MAXIMUM_CHUNK_SIZE))
-
-  /**
-   * Build a `reply_channel_range` message
-   *
-   * @param chunk           chunk of scids
-   * @param chainHash       chain hash
-   * @param defaultEncoding default encoding
-   * @param queryFlags_opt  query flag set by the requester
-   * @param channels        channels map
-   * @return a ReplyChannelRange object
-   */
-  def buildReplyChannelRange(chunk: ShortChannelIdsChunk, chainHash: ByteVector32, defaultEncoding: EncodingType, queryFlags_opt: Option[QueryChannelRangeTlv.QueryFlags], channels: SortedMap[ShortChannelId, PublicChannel]): ReplyChannelRange = {
-    val encoding = if (chunk.shortChannelIds.isEmpty) EncodingType.UNCOMPRESSED else defaultEncoding
-    val (timestamps, checksums) = queryFlags_opt match {
-      case Some(extension) if extension.wantChecksums | extension.wantTimestamps =>
-        // we always compute timestamps and checksums even if we don't need both, overhead is negligible
-        val (timestamps, checksums) = chunk.shortChannelIds.map(getChannelDigestInfo(channels)).unzip
-        val encodedTimestamps = if (extension.wantTimestamps) Some(ReplyChannelRangeTlv.EncodedTimestamps(encoding, timestamps)) else None
-        val encodedChecksums = if (extension.wantChecksums) Some(ReplyChannelRangeTlv.EncodedChecksums(checksums)) else None
-        (encodedTimestamps, encodedChecksums)
-      case _ => (None, None)
-    }
-    ReplyChannelRange(chainHash, chunk.firstBlock, chunk.numBlocks,
-      complete = 1,
-      shortChannelIds = EncodedShortChannelIds(encoding, chunk.shortChannelIds),
-      timestamps = timestamps,
-      checksums = checksums)
-  }
-
-  def addToSync(syncMap: Map[PublicKey, Sync], remoteNodeId: PublicKey, pending: List[RoutingMessage]): (Map[PublicKey, Sync], Option[RoutingMessage]) = {
-    pending match {
-      case head +: rest =>
-        // they may send back several reply_channel_range messages for a single query_channel_range query, and we must not
-        // send another query_short_channel_ids query if they're still processing one
-        syncMap.get(remoteNodeId) match {
-          case None =>
-            // we don't have a pending query with this peer, let's send it
-            (syncMap + (remoteNodeId -> Sync(rest, pending.size)), Some(head))
-          case Some(sync) =>
-            // we already have a pending query with this peer, add missing ids to our "sync" state
-            (syncMap + (remoteNodeId -> Sync(sync.pending ++ pending, sync.total + pending.size)), None)
-        }
-      case Nil =>
-        // there is nothing to send
-        (syncMap, None)
-    }
-  }
-
-  /**
-   * https://github.com/lightningnetwork/lightning-rfc/blob/master/04-onion-routing.md#clarifications
-   */
-  val ROUTE_MAX_LENGTH = 20
-
-  // Max allowed CLTV for a route
-  val DEFAULT_ROUTE_MAX_CLTV = CltvExpiryDelta(1008)
-
-  // The default number of routes we'll search for when findRoute is called with randomize = true
-  val DEFAULT_ROUTES_COUNT = 3
-
-  def getDefaultRouteParams(routerConf: RouterConf) = RouteParams(
-    randomize = routerConf.randomizeRouteSelection,
-    maxFeeBase = routerConf.searchMaxFeeBase.toMilliSatoshi,
-    maxFeePct = routerConf.searchMaxFeePct,
-    routeMaxLength = routerConf.searchMaxRouteLength,
-    routeMaxCltv = routerConf.searchMaxCltv,
-    ratios = routerConf.searchHeuristicsEnabled match {
-      case false => None
-      case true => Some(WeightRatios(
-        cltvDeltaFactor = routerConf.searchRatioCltv,
-        ageFactor = routerConf.searchRatioChannelAge,
-        capacityFactor = routerConf.searchRatioChannelCapacity
-      ))
-    }
-  )
-
-  /**
-   * Find a route in the graph between localNodeId and targetNodeId, returns the route.
-   * Will perform a k-shortest path selection given the @param numRoutes and randomly select one of the result.
-   *
-   * @param g            graph of the whole network
-   * @param localNodeId  sender node (payer)
-   * @param targetNodeId target node (final recipient)
-   * @param amount       the amount that will be sent along this route
-   * @param numRoutes    the number of shortest-paths to find
-   * @param extraEdges   a set of extra edges we want to CONSIDER during the search
-   * @param ignoredEdges a set of extra edges we want to IGNORE during the search
-   * @param routeParams  a set of parameters that can restrict the route search
-   * @return the computed route to the destination @targetNodeId
-   */
-  def findRoute(g: DirectedGraph,
-                localNodeId: PublicKey,
-                targetNodeId: PublicKey,
-                amount: MilliSatoshi,
-                numRoutes: Int,
-                extraEdges: Set[GraphEdge] = Set.empty,
-                ignoredEdges: Set[ChannelDesc] = Set.empty,
-                ignoredVertices: Set[PublicKey] = Set.empty,
-                routeParams: RouteParams,
-                currentBlockHeight: Long): Try[Seq[ChannelHop]] = Try {
-
-    if (localNodeId == targetNodeId) throw CannotRouteToSelf
-
-    def feeBaseOk(fee: MilliSatoshi): Boolean = fee <= routeParams.maxFeeBase
-
-    def feePctOk(fee: MilliSatoshi, amount: MilliSatoshi): Boolean = {
-      val maxFee = amount * routeParams.maxFeePct
-      fee <= maxFee
-    }
-
-    def feeOk(fee: MilliSatoshi, amount: MilliSatoshi): Boolean = feeBaseOk(fee) || feePctOk(fee, amount)
-
-    def lengthOk(length: Int): Boolean = length <= routeParams.routeMaxLength && length <= ROUTE_MAX_LENGTH
-
-    def cltvOk(cltv: CltvExpiryDelta): Boolean = cltv <= routeParams.routeMaxCltv
-
-    val boundaries: RichWeight => Boolean = { weight =>
-      feeOk(weight.cost - amount, amount) && lengthOk(weight.length) && cltvOk(weight.cltv)
-    }
-
-    val foundRoutes = KamonExt.time(Metrics.FindRouteDuration.withTag(Tags.NumberOfRoutes, numRoutes).withTag(Tags.Amount, Tags.amountBucket(amount))) {
-      Graph.yenKshortestPaths(g, localNodeId, targetNodeId, amount, ignoredEdges, ignoredVertices, extraEdges, numRoutes, routeParams.ratios, currentBlockHeight, boundaries).toList
-    }
-    foundRoutes match {
-      case Nil if routeParams.routeMaxLength < ROUTE_MAX_LENGTH => // if not found within the constraints we relax and repeat the search
-        Metrics.RouteLength.withTag(Tags.Amount, Tags.amountBucket(amount)).record(0)
-        return findRoute(g, localNodeId, targetNodeId, amount, numRoutes, extraEdges, ignoredEdges, ignoredVertices, routeParams.copy(routeMaxLength = ROUTE_MAX_LENGTH, routeMaxCltv = DEFAULT_ROUTE_MAX_CLTV), currentBlockHeight)
-      case Nil =>
-        Metrics.RouteLength.withTag(Tags.Amount, Tags.amountBucket(amount)).record(0)
-        throw RouteNotFound
-      case foundRoutes =>
-        val routes = foundRoutes.find(_.path.size == 1) match {
-          case Some(directRoute) => directRoute :: Nil
-          case _ => foundRoutes
-        }
-        // At this point 'routes' cannot be empty
-        val randomizedRoutes = if (routeParams.randomize) Random.shuffle(routes) else routes
-        val route = randomizedRoutes.head.path.map(graphEdgeToHop)
-        Metrics.RouteLength.withTag(Tags.Amount, Tags.amountBucket(amount)).record(route.length)
-        route
-    }
-  }
 }
