+++ eclair-core/src/test/scala/fr/acinq/eclair/integration/IntegrationSpec.scala
@@ -46,7 +46,7 @@ import fr.acinq.eclair.payment.relay.Relayer.{GetOutgoingChannels, OutgoingChann
 import fr.acinq.eclair.payment.send.PaymentInitiator.{SendPaymentRequest, SendTrampolinePaymentRequest}
 import fr.acinq.eclair.payment.send.PaymentLifecycle.{State => _}
 import fr.acinq.eclair.router.Graph.WeightRatios
-import fr.acinq.eclair.router.Router.ROUTE_MAX_LENGTH
+import fr.acinq.eclair.router.RouteCalculationHandlers.ROUTE_MAX_LENGTH
 import fr.acinq.eclair.router.{NORMAL => _, State => _, _}
 import fr.acinq.eclair.transactions.Transactions
 import fr.acinq.eclair.transactions.Transactions.{HtlcSuccessTx, HtlcTimeoutTx}
